+++ test/test-setup.js
@@ -1,7 +1,9 @@
 require('babel-register');
 const Enzyme = require('enzyme');
 const Adapter = require('enzyme-adapter-react-16');
-Enzyme.configure({ adapter: new Adapter() });
+Enzyme.configure({
+  adapter: new Adapter(),
+});
 const jsdom = require('jsdom').jsdom;
 
 global.document = jsdom('');