+++ CHANGELOG.md
@@ -2,13 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-<a name="3.2.0"></a>
-# [3.2.0](https://github.com/mongodb-js/connect-backbone-to-react/compare/v3.0.0...v3.2.0) (2020-03-06)
-
 
 
 <a name="3.1.0"></a>
 # [3.1.0](https://github.com/mongodb-js/connect-backbone-to-react/compare/v3.0.0...v3.1.0) (2020-03-06)
+
+* Adds the useBackboneContext hook, which can be called to return the models and collections converted to JSON for usage in functional components